Fixed minor bug in Writer
The grid truncation feature was being overzealous and truncating the
cutoff to be an integer (but only for the header line!). Fixed.

+++ src/writer/Writer.cpp
@@ -86,7 +86,7 @@ void Writer::writeVertical(bool reset) {
 }
 
 
-void Writer::writeHeader(int cutoff) {
+void Writer::writeHeader(float cutoff) {
   ofstream file(conf->output_file->path.string());
   time_t now = chrono::system_clock::to_time_t(chrono::system_clock::now());
   string date = ctime(&now);
